use intptr_t as offset_t again

+++ include/cista/offset_t.h
@@ -7,9 +7,7 @@
 
 namespace cista {
 
-// using offset_t = intptr_t;
-
-using offset_t = int64_t;
+using offset_t = intptr_t;
 
 constexpr auto const NULLPTR_OFFSET = std::numeric_limits<offset_t>::min();
 constexpr auto const DANGLING = std::numeric_limits<offset_t>::min() + 1;
